refactor: move tests for pub functions
