--- tests/application/course_schedule.rs
@@ -1,1 +1,0 @@
-
